full gpu - testing

+++ NPB-RAYON/src/cg.rs
@@ -1012,7 +1012,7 @@ mod cg {
     fn alloc_a_d() -> Vec<f64> { vec![0.0; NZ] }
     #[kernelversion(acc_count=(AtLeast{val:1}), acc_backend=CUDA)]
     fn alloc_a_d() -> Vec<f64> { 
-        let mut ptr: *mut i32 = std::ptr::null_mut();
+        let mut ptr: *mut f64 = std::ptr::null_mut();
         unsafe { alloc_a_gpu(&mut ptr, NZ as i32) };
         unsafe { Vec::from_raw_parts(ptr, NZ, NZ) };
     }
@@ -1059,7 +1059,7 @@ mod cg {
     fn alloc_x() -> Vec<f64> { vec![1.0; NA as usize + 2] }
     #[kernelversion(acc_count=(AtLeast{val:1}), acc_backend=CUDA)]
     fn alloc_x() -> Vec<f64> { 
-        let mut ptr: *mut i32 = std::ptr::null_mut();
+        let mut ptr: *mut f64 = std::ptr::null_mut();
         unsafe { alloc_x_gpu(&mut ptr, NA + 2) };
         unsafe { Vec::from_raw_parts(ptr, (NA + 2) as usize, (NA + 2) as usize) };
     }
@@ -1070,7 +1070,7 @@ mod cg {
     fn alloc_z() -> Vec<f64> { vec![0.0; NA as usize + 2] }
     #[kernelversion(acc_count=(AtLeast{val:1}), acc_backend=CUDA)]
     fn alloc_z() -> Vec<f64> { 
-        let mut ptr: *mut i32 = std::ptr::null_mut();
+        let mut ptr: *mut f64 = std::ptr::null_mut();
         unsafe { alloc_z_gpu(&mut ptr, NA + 2) };
         unsafe { Vec::from_raw_parts(ptr, (NA + 2) as usize, (NA + 2) as usize) };
     }
@@ -1081,7 +1081,7 @@ mod cg {
     fn alloc_p() -> Vec<f64> { vec![0.0; NA as usize + 2] }
     #[kernelversion(acc_count=(AtLeast{val:1}), acc_backend=CUDA)]
     fn alloc_p() -> Vec<f64> { 
-        let mut ptr: *mut i32 = std::ptr::null_mut();
+        let mut ptr: *mut f64 = std::ptr::null_mut();
         unsafe { alloc_p_gpu(&mut ptr, NA + 2) };
         unsafe { Vec::from_raw_parts(ptr, (NA + 2) as usize, (NA + 2) as usize) };
     }
@@ -1092,7 +1092,7 @@ mod cg {
     fn alloc_q() -> Vec<f64> { vec![0.0; NA as usize + 2] }
     #[kernelversion(acc_count=(AtLeast{val:1}), acc_backend=CUDA)]
     fn alloc_q() -> Vec<f64> { 
-        let mut ptr: *mut i32 = std::ptr::null_mut();
+        let mut ptr: *mut f64 = std::ptr::null_mut();
         unsafe { alloc_q_gpu(&mut ptr, NA + 2) };
         unsafe { Vec::from_raw_parts(ptr, (NA + 2) as usize, (NA + 2) as usize) };
     }
@@ -1103,7 +1103,7 @@ mod cg {
     fn alloc_r() -> Vec<f64> { vec![0.0; NA as usize + 2] }
     #[kernelversion(acc_count=(AtLeast{val:1}), acc_backend=CUDA)]
     fn alloc_r() -> Vec<f64> { 
-        let mut ptr: *mut i32 = std::ptr::null_mut();
+        let mut ptr: *mut f64 = std::ptr::null_mut();
         unsafe { alloc_r_gpu(&mut ptr, NA + 2) };
         unsafe { Vec::from_raw_parts(ptr, (NA + 2) as usize, (NA + 2) as usize) };
     }
